+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - CSS Classes for Radio/Checkbox containers
+- Mustache rendering to Info Elements
 
 ## [2.6.0] - 2023-02-16